backend/requirements.txt
Django==2.2.14
psycopg2
pytz
sqlparse
Pillow
djangorestframework
django-cors-headers
django-allauth
dj-rest-auth